Updated parser and makefile
Parser can interpret statements of the form "x_1+x_2+...+x_n;" now. This is just a quick upload to make sure I have a copy of this because it's a good framework for the state machine to check a grammar and apply a semantic rule. Also added the debug option in the makefile and made sure it recompiled on chages to all major files (driver.c, parser.c, lexer.c). I might update the lexer so that math token's token is just their attribute (+, -, *, /, etc) because it will make some parts easier to write and will also help with expressions like "*ptr = 5;" and "ident=*ptr+7;" because now instead of the token being MATH and the attribute being '*', the token itself will just be '*' which is more general.